+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- [#287] Added PublishedValue class
 
 ### Fixed
 - [#283] Cached proxy objects are visible after exiting play mode